requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
requests
pydantic
jinja2
matplotlib
scipy
tqdm
pyyaml
biopython
